Add authentication and file management commands
- Implement `check_auth` and `add_token` functions for token management in `auth.rs`.
- Create `move_files` function for moving files with progress tracking in `file_ops.rs`.
- Add `open_resource_file`, `copy_premiere_project`, and `show_confirmation_dialog` in `premiere.rs`.
- Introduce `get_folders` and `upload_video` functions for video uploads in `sprout_upload.rs`.
- Implement `graceful_restart` and `get_username` commands in `system.rs`.
- Add `copy_file_with_progress` utility function for file copying with progress updates.
- Update module declarations in `mod.rs` files to include new commands and utilities.
