+++ src-tauri/src/state/auth.rs
@@ -1,0 +1,5 @@
+use std::sync::Mutex;
+
+pub struct AuthState {
+    pub tokens: Mutex<Vec<String>>, // Simple token storage
+}
