+++ src-tauri/src/state/mod.rs
@@ -1,0 +1,3 @@
+pub mod auth;
+
+pub use auth::*;
